feat(home): :sparkles: implement HomeActivity with edge-to-edge layout and control buttons
